Move Console tests out of the Unit Tests directory
This makes it clear that unit tests should not be starting the Application.
