+++ app/phpunit.xml
@@ -7,6 +7,9 @@
     <testsuite name="Unit Tests">
       <directory suffix="Test.php">./tests/Unit</directory>
     </testsuite>
+    <testsuite name="Console Tests">
+      <directory suffix="Test.php">./tests/Console</directory>
+    </testsuite>
   </testsuites>
   <php>
     <env name="APP_ENV" value="testing"/>
