+++ app/tests/Unit/ExampleTest.php
@@ -2,7 +2,7 @@
 
 namespace Tests\Unit;
 
-use Tests\TestCase;
+use PHPUnit\Framework\TestCase;
 
 /**
  * @internal
